refactor: relocate startup tests and middleware test
