+++ internal/app/run.go
@@ -7,11 +7,9 @@ import (
 	"log"
 	"net/http"
 	"os"
-	"path/filepath"
-	"strings"
 	"time"
 
-	goa4web "github.com/arran4/goa4web"
+	"github.com/arran4/goa4web"
 	"github.com/arran4/goa4web/config"
 	"github.com/arran4/goa4web/core"
 	common "github.com/arran4/goa4web/core/common"
@@ -64,7 +62,7 @@ func RunWithConfig(ctx context.Context, cfg runtimeconfig.RuntimeConfig, session
 		return fmt.Errorf("startup checks: %w", err)
 	}
 
-	dbPool := goa4web.GetDBPool()
+	dbPool := GetDBPool()
 	if err := corelanguage.ValidateDefaultLanguage(context.Background(), goa4web.New(dbPool), cfg.DefaultLanguage); err != nil {
 		return fmt.Errorf("default language: %w", err)
 	}
@@ -144,33 +142,3 @@ type routerWrapper interface {
 type routerWrapperFunc func(http.Handler) http.Handler
 
 func (f routerWrapperFunc) Wrap(h http.Handler) http.Handler { return f(h) }
-
-func performStartupChecks(cfg runtimeconfig.RuntimeConfig) error {
-	if ue := goa4web.InitDB(cfg); ue != nil {
-		return fmt.Errorf("%s: %w", ue.ErrorMessage, ue.Err)
-	}
-	if err := checkUploadDir(cfg); err != nil {
-		return err
-	}
-	return nil
-}
-
-func checkUploadDir(cfg runtimeconfig.RuntimeConfig) error {
-	if cfg.ImageUploadDir == "" {
-		return fmt.Errorf("image upload directory not set")
-	}
-	if strings.HasPrefix(cfg.ImageUploadDir, "s3://") {
-		// TODO: validate S3 upload targets
-		return nil
-	}
-	info, err := os.Stat(cfg.ImageUploadDir)
-	if err != nil || !info.IsDir() {
-		return fmt.Errorf("image upload directory invalid: %w", err)
-	}
-	test := filepath.Join(cfg.ImageUploadDir, ".check")
-	if err := os.WriteFile(test, []byte("ok"), 0644); err != nil {
-		return fmt.Errorf("image upload directory not writable: %w", err)
-	}
-	os.Remove(test)
-	return nil
-}
